archi rework : moved payload validation logic to a dedicated package
