+++ go.mod
@@ -4,10 +4,10 @@ go 1.18
 
 require (
 	cuelang.org/go v0.4.2
-	github.com/fsnotify/fsnotify v1.5.1
 	github.com/labstack/echo/v4 v4.7.2
 	github.com/perses/common v0.10.0
 	github.com/perses/perses v0.0.0-20220323190413-5567c3dc05e1
 	github.com/prometheus/common v0.32.1
 	github.com/sirupsen/logrus v1.8.1
+	gopkg.in/fsnotify.v1 v1.4.7
 )
